Some changes

+++ Application/Domain/User.cs
@@ -2,7 +2,7 @@ using Application.Domain;
 
 public class User
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; }
     public string Email { get; set; }
     public string PasswordHash { get; private set; }
@@ -25,7 +25,7 @@ public class User
         DateOfBirth = dateOfBirth;
     }
 
-    public User(long id, string name, string email, string passwordHash, DateOnly dateOfBirth)
+    public User(int id, string name, string email, string passwordHash, DateOnly dateOfBirth)
     {
         Id = id;
         Name = name;
